+++ lib/src/domain/analyzer/use_case/get_exit_code.dart
@@ -4,11 +4,13 @@ import 'package:pull_request_coverage/src/domain/user_options/models/user_option
 
 class GetExitCode {
   int call(AnalysisResult analysisResult, UserOptions userOptions) {
-    if (userOptions.minimumCoverageRate != null && analysisResult.coverageRate < (userOptions.minimumCoverageRate! / 100)) {
+    if (userOptions.minimumCoverageRate != null &&
+        analysisResult.coverageRate < (userOptions.minimumCoverageRate! / 100)) {
       return ExitCode.testFail;
     }
 
-    if (userOptions.maximumUncoveredLines != null && analysisResult.totalOfUncoveredNewLines > userOptions.maximumUncoveredLines!) {
+    if (userOptions.maximumUncoveredLines != null &&
+        analysisResult.linesMissingTests > userOptions.maximumUncoveredLines!) {
       return ExitCode.testFail;
     }