+++ lib/src/domain/input_reader/diff_reader/models/file_diff.dart
@@ -9,6 +9,7 @@ class FileDiff {
 
   int get newLinesCount => lines.where((line) => line.isANewLine).length;
   int get uncoveredNewLinesCount => lines.where((line) => line.isTestMissing).length;
+  int get ignoredUntestedLinesCount => lines.where((line) => line.ignored == true && line.isUntested == true && line.isANewLine).length;
 
   const FileDiff({
     required this.path,