+++ lib/src/domain/input_reader/diff_reader/models/file_line.dart
@@ -1,4 +1,4 @@
-/// [isUncovered] is `false` when `FileLine` is created by `ParseGitDiff`
+/// [isUntested] is `false` when `FileLine` is created by `ParseGitDiff`
 /// To coverage will be analysed by `Analyze` use case and this value will be updated there.
 ///
 /// [isANewLine] tells if this line was added in the current PR.
@@ -6,20 +6,20 @@ class FileLine {
   final String line;
   final int lineNumber;
   final bool isANewLine;
-  bool? isUncovered;
+  bool? isUntested;
   bool? ignored;
 
   FileLine({
     required this.line,
     required this.lineNumber,
     required this.isANewLine,
-    this.isUncovered,
+    this.isUntested,
     this.ignored,
   });
 
   bool get isANewNotIgnoredLine => isANewLine && ignored == false;
 
-  bool get isTestMissing => isANewLine && isUncovered == true && ignored == false;
+  bool get isTestMissing => isANewLine && isUntested == true && ignored == false;
 
   @override
   String toString() => "$lineNumber ${isTestMissing ? "💃" : "👍"}: $line";