+++ lib/src/domain/user_options/models/user_options.dart
@@ -3,13 +3,15 @@ import 'package:pull_request_coverage/src/domain/user_options/models/markdown_mo
 import 'package:pull_request_coverage/src/domain/user_options/models/output_mode.dart';
 
 class UserOptions {
-  final List<Glob> fileFilters;
+  final List<Glob> ignoredFiles;
+  final List<Glob> knownGeneratedFiles;
   final String lcovFilePath;
   final String? fullyTestedMessage;
   final double? minimumCoverageRate;
   final int? maximumUncoveredLines;
   final bool showUncoveredCode;
   final bool useColorfulOutput;
+  final bool ignoreKnownGeneratedFiles;
   final bool reportFullyCoveredFiles;
   final OutputMode outputMode;
   final MarkdownMode markdownMode;
@@ -20,16 +22,18 @@ class UserOptions {
 
   /// [fractionalDigits] how many digits after the decimal point to show on coverage rate
   const UserOptions({
-    required this.fileFilters,
+    required this.ignoredFiles,
     required this.lcovFilePath,
     required this.showUncoveredCode,
     required this.outputMode,
     required this.fractionalDigits,
     required this.markdownMode,
     this.fullyTestedMessage,
+    this.knownGeneratedFiles = const [],
     this.minimumCoverageRate,
     this.maximumUncoveredLines,
     this.useColorfulOutput = true,
+    this.ignoreKnownGeneratedFiles = true,
     this.reportFullyCoveredFiles = true,
     this.stdinTimeout = const Duration(seconds: 1),
     this.deprecatedFilterSet = false,